refactor(thread): make own package, refactor test_compile.sh
