+++ cmd/photos.go
@@ -7,9 +7,9 @@ import (
 	"github.com/fatih/color"
 	"github.com/mitchellh/go-homedir"
 	"github.com/textileio/textile-go/core"
-	"github.com/textileio/textile-go/core/thread"
 	"github.com/textileio/textile-go/ipfs"
 	"github.com/textileio/textile-go/repo"
+	"github.com/textileio/textile-go/thread"
 	"gopkg.in/abiosoft/ishell.v2"
 	"io/ioutil"
 	"os"
