+++ core/photos.go
@@ -6,10 +6,10 @@ import (
 	"errors"
 	"fmt"
 	"github.com/textileio/textile-go/archive"
-	"github.com/textileio/textile-go/core/thread"
 	"github.com/textileio/textile-go/crypto"
 	"github.com/textileio/textile-go/ipfs"
 	"github.com/textileio/textile-go/photo"
+	"github.com/textileio/textile-go/thread"
 	uio "gx/ipfs/Qmb8jW1F6ZVyYPW1epc2GFRipmd3S8tJ48pZKBVPzVqj9T/go-ipfs/unixfs/io"
 	"os"
 	"path/filepath"
