+++ thread/helpers.go
