+++ thread/ignores.go
