+++ thread/main.go
