+++ thread/main_test.go
@@ -1,0 +1,7 @@
+package thread_test
+
+import "testing"
+
+func Test_SetupThread(t *testing.T) {
+
+}
